Ability to search for text in mails
Finding substrings in the displayed mail body is something users do. We
could pipe it to less and search there, but that is a very cumbersome
approach.

This patch adds the ability to search the mail body. This comes with a
cost: we are loosing the ability to render the mailbody with a variable
length. In other words: the renderer now uses hard line breaks in order
to calculate the position of the highlighted words.